cloud_functions/process_weather_data_function/requirements.txt
requests
google-cloud-bigquery
flask
datetime
google-cloud-storage
polars